+++ app/Filament/Resources/ServiceCategories/ServiceCategoryResource.php
@@ -3,6 +3,7 @@
 namespace App\Filament\Resources\ServiceCategories;
 
 use App\Models\ServiceCategory;
+use BackedEnum;
 use Filament\Actions\DeleteAction;
 use Filament\Actions\EditAction;
 use Filament\Forms\Components\Textarea;
@@ -10,16 +11,22 @@ use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\Toggle;
 use Filament\Resources\Resource;
 use Filament\Schemas\Schema;
+use Filament\Support\Icons\Heroicon;
 use Filament\Tables\Columns\IconColumn;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Filters\TernaryFilter;
 use Filament\Tables\Table;
+use UnitEnum;
 
 class ServiceCategoryResource extends Resource
 {
     protected static ?string $model = ServiceCategory::class;
 
     protected static ?string $recordTitleAttribute = 'name';
+
+    protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedTag;
+
+    protected static string|UnitEnum|null $navigationGroup = 'Service Management';
 
     public static function form(Schema $schema): Schema
     {