+++ app/Models/SystemBackup.php
@@ -11,7 +11,7 @@ class SystemBackup extends Model
 {
     public const UPDATED_AT = null;
 
-    protected $guarded = ['*'];
+    protected $guarded = [];
 
     protected function casts(): array
     {